database/fut_pulse/requirements.txt
pyautogui
Pillow
cnocr
numpy
onnxruntime
pymysql
python-dotenv
akshare
